Implement inverse dupe map for sequence. Strip all terms on sequence dupe match.  Behavioral change, will up semver. Additional unit test coverage.
Main goal of this PR is to strip all terms on a match with duplicates,
not just the first term of the dupe range was was the previous behavior.
This new behavior is more consistent with sets and is easier to
understand. It also removes more terms out of the search space which
improves performance.

To accomplish this, I modified the inverse sequence implementation to
use a dupeMap instead of a bit mask to account for dupes. This is more
in keeping with the other implementations and removes the maxTerms of
64 restriction on dupes.